Test for comment (#6)
* feat: test all controllers ok

* add all test with ui in terminal

* add all test with ui in terminal

+++ controllers/comment/delete.js
@@ -2,7 +2,7 @@ const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 
 exports.delete = async (req, res) => {
-  console.log("delete comment route", req)
+  console.log("delete comment route", req);
   try {
     const commentId = parseInt(req.params.commentId);
 
@@ -22,14 +22,16 @@ exports.delete = async (req, res) => {
       },
     });
 
+    if (!commentToDelet) {
+      return res.status(404).json({ error: "Comment not found" });
+    }
+
     if (isAdmin || commentToDelet.userId === userIdFromToken) {
       const deletedComment = await prisma.comment.delete({
         where: { id: commentId },
       });
 
-      res
-        .status(200)
-        .json({ message: "Comment deleted", data: deletedComment });
+      res.status(200).json({ message: "Comment deleted", data: deletedComment });
     } else {
       res.status(403).json({ error: "Unauthorized" });
     }
@@ -38,11 +40,8 @@ exports.delete = async (req, res) => {
       "Error deleting comment:",
       error.message || "Internal Server Error"
     );
-    res
-      .status(500)
-      .json({ error: "Error deleting comment", details: error.message });
+    res.status(500).json({ error: "Error deleting comment", details: error.message });
   } finally {
-    console.log(res)
     await prisma.$disconnect();
   }
 };
